Create serverless function for aggregated counts
The main PubMed search page at https://pubmed.ncbi.nlm.nih.gov/ has some
backend logic that aggregates publication count by year. This allows us
to extract all of the aggregated data in a single request, rather than a
single API request for each year, which seems to be all that the
eutilities API allows. It's not the prettiest solution, but it performs
like a rocket.